+++ tests/test.sh
@@ -6,4 +6,4 @@ python3 test_mincap.py
 python3 test_reachability.py
 python3 test_safety.py
 python3 test_strategy.py
-
+jupyter nbconvert --execute ../tut/StrategyTypes.ipynb