+++ tests/test.sh
@@ -6,5 +6,6 @@ python3 test_mecs.py
 python3 test_mincap.py
 python3 test_reachability.py
 python3 test_safety.py
+python3 test_strategy.py
 python3 test_strategy_old.py
 jupyter nbconvert --execute ../tut/StrategyTypes.ipynb